third_party/WebKit/LayoutTests/http/tests/security/w3c/cross-origin-objects-expected.txt

This is a testharness.js-based test.
PASS Basic sanity-checking 
FAIL Only whitelisted properties are accessible cross-origin Blocked a frame with origin "http://web-platform.test:8000" from accessing a cross-origin frame.
PASS [[GetPrototypeOf]] should return null 
PASS [[SetPrototypeOf]] should throw 
PASS [[IsExtensible]] should return true for cross-origin objects 
PASS [[PreventExtensions]] should throw for cross-origin objects 
FAIL [[GetOwnProperty]] - Properties on cross-origin objects should be reported |own| Blocked a frame with origin "http://web-platform.test:8000" from accessing a cross-origin frame.
FAIL [[GetOwnProperty]] - Property descriptors for cross-origin properties should be set up correctly assert_equals: property descriptor for location should be non-enumerable expected false but got true
PASS [[Delete]] Should throw on cross-origin objects 
FAIL [[DefineOwnProperty]] Should throw for cross-origin objects assert_throws: Can't define cross-origin value property length function "function () { Object.defineProperty(obj, prop, valueDesc); }" did not throw
PASS [[Enumerate]] should return an empty iterator 
FAIL [[OwnPropertyKeys]] should return all properties from cross-origin objects assert_array_equals: Object.getOwnPropertyNames() gives the right answer for cross-origin Location property 0, expected "href" but got "assign"
PASS A and B jointly observe the same identity for cross-origin Window and Location 
PASS Cross-origin functions get local Function.prototype 
FAIL Cross-origin Window accessors get local Function.prototype Cannot read property 'name' of undefined
FAIL Same-origin observers get different functions for cross-origin objects assert_true: same-origin Window functions get their own object expected true got false
FAIL Same-origin obsevers get different accessors for cross-origin Window assert_true: different Window accessors per-incumbent script settings object expected true got false
FAIL Same-origin observers get different accessors for cross-origin Location Blocked a frame with origin "http://web-platform.test:8000" from accessing a cross-origin frame.
PASS Cross-origin object identity preserved across document.domain 
Harness: the test ran to completion.
